pyproject.toml
[project]
name = "maplediscord"
version = "0.1.0"
description = ""
authors = [
    {name = "sumin",email = "ksum.jerry@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "supabase (>=2.17.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
